requirements.txt
attrs==20.3.0
beautifulsoup4==4.9.3
blinker==1.4
Brlapi==0.8.2
cached-property==1.5.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.5.0
docker==4.4.4
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
evdev==1.4.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
hidpidaemon==18.4.6
html5lib==1.1
httplib2==0.18.1
idna==2.10
IMDbPY==6.8
importlib-metadata==4.0.1
itsdangerous==2.0.1
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
kernelstub==3.1.4
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
louis==3.18.0
lxml==4.6.3
macaroonbakery==1.3.1
MarkupSafe==2.0.1
more-itertools==4.2.0
mysql-connector==2.2.9
natsort==7.1.0
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
Pillow==8.1.2
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.16.2
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.40.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt===2.2.1pop0-1628521831-21.10-5dbe6cf
python-debian===0.1.39ubuntu1
python-dotenv==0.15.0
python-magic==0.4.20
python-xlib==0.29
pytz==2021.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.3.1
repolib==1.5.2
repoman==1.4.0
requests==2.25.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
setproctitle==1.2.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
systemd-python==234
texttable==1.6.3
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.2
wxPython==4.0.7
xdg==5
xkit==0.0.0
youtube-dl==2021.6.6
zipp==1.0.0
